renamed process_context to context_tools
